chore: move test folder and rename it to __test__ to include it without compilation issues
