+++ __mocks__/@zextras/carbonio-shell-ui.ts
@@ -6,7 +6,7 @@
 
 import shell from '@zextras/carbonio-shell-ui';
 
-export * from '../../src/test/mocks/carbonio-shell-ui';
+export * from '../../src/__test__/mocks/carbonio-shell-ui';
 
 // TODO move it in the Commons submodule
 export const useAuthenticated = jest
