+++ src/__test__/mocks/utils/window.ts
